# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Users microservice",
        "title": "Go Microservice Users API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/{id}/cloak": {
            "get": {
                "description": "Получить URL плаща пользователя по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Получить URL плаща пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Загружает плащ пользователя по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить плащ пользователю",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл плаща (png)",
                        "name": "cloak",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_UploadResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка загрузки или открытия файла",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет плащ пользователя по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Удалить плащ пользоваетя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_Response"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/skin": {
            "get": {
                "description": "Получить URL скина пользователя по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Получить URL скина пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Загружает скин пользователя по его ID. Вырезает голову из скина и загружает в avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить скин пользователю",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл скина (png)",
                        "name": "skin",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_UploadResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка загрузки файла",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет скин пользователя по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Удалить скин пользоваетя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/ginx.SuccessResponse-internal_infra_http_controllers_files_Response"
                        }
                    },
                    "404": {
                        "description": "Пользователя не существует",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления",
                        "schema": {
                            "$ref": "#/definitions/ginx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ginx.ErrorResponse": {
            "description": "Стандартная структура ответа при возникновении ошибки",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Код ошибки для клиента",
                    "type": "string",
                    "example": "ERROR_CODE"
                },
                "details": {
                    "description": "Дополнительные детали ошибки",
                    "type": "string",
                    "example": "Дополнительные детали ошибки"
                },
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Произошла ошибка"
                }
            }
        },
        "ginx.SuccessResponse-internal_infra_http_controllers_files_Response": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_infra_http_controllers_files.Response"
                },
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Операция выполнена успешно"
                }
            }
        },
        "ginx.SuccessResponse-internal_infra_http_controllers_files_UploadResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_infra_http_controllers_files.UploadResponse"
                },
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Операция выполнена успешно"
                }
            }
        },
...
        "internal_infra_http_controllers_files.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_infra_http_controllers_files.UploadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Тип: Bearer {token}. Токен получается через Keycloak.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}